requirements.txt
streamlit
google-genai
python-dotenv
pandas
matplotlib
seaborn
numpy